+++ src/main.rs
@@ -2,43 +2,57 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-use bevy::{
-    log::{Level, LogPlugin},
-    prelude::*,
-    window::WindowTheme,
-};
+use bevy::{log::LogPlugin, prelude::*, window::WindowTheme};
 
 use crate::{
-    action::ActionPlugin, camera::CameraPlugin, cell::CellPlugin, grid::GridPlugin,
-    input::InputPlugin,
+    action::ActionPlugin, app::AppPlugin, camera::CameraPlugin, cell::CellPlugin,
+    feature::FeaturePlugin, grid::GridPlugin, input::InputPlugin, inspect::InspectPlugin,
+    schedule::SchedulePlugin, ui::UiPlugin,
 };
 
 mod action;
+mod app;
 mod camera;
 mod cell;
+mod feature;
 mod grid;
 mod input;
+mod schedule;
+mod ui;
+
+#[cfg(feature = "inspect")]
+mod inspect;
+
+#[cfg(not(feature = "inspect"))]
+mod inspect {
+    use crate::feature::FeaturePlugin;
+
+    pub struct InspectPlugin;
+    impl FeaturePlugin for InspectPlugin {}
+}
 
 fn main() {
-    App::new()
-        .add_plugins((DefaultPlugins
+    let mut app = App::new();
+    app.add_plugins(
+        DefaultPlugins
             .set(WindowPlugin {
                 primary_window: Some(Window {
                     title: "hanjie".to_string(),
                     window_theme: Some(WindowTheme::Dark),
-                    ..Default::default()
+                    ..default()
                 }),
-                ..Default::default()
+                ..default()
             })
-            .set(LogPlugin {
-                filter: "info,wgpu_core=warn,wgpu_hal=warn,hanjie=debug".into(),
-                level: Level::DEBUG,
-            }),))
-        .add_plugins(ActionPlugin)
-        .add_plugins(CameraPlugin)
-        .add_plugins(CellPlugin)
-        .add_plugins(GridPlugin)
-        .add_plugins(InputPlugin)
-        .add_systems(Update, bevy::window::close_on_esc)
-        .run();
+            .disable::<LogPlugin>(),
+    )
+    .add_plugins(ActionPlugin)
+    .add_plugins(AppPlugin)
+    .add_plugins(CameraPlugin)
+    .add_plugins(CellPlugin)
+    .add_plugins(GridPlugin)
+    .add_plugins(InputPlugin)
+    .add_plugins(SchedulePlugin)
+    .add_plugins(UiPlugin);
+    InspectPlugin::load(&mut app);
+    app.run();
 }